requirements.txt
uvicorn
requests
sentence-transformers
elasticsearch==8.12.0
tqdm
fastapi
pydantic
python-multipart
pandas
kafka-python
aiokafka==0.10.0
pydantic-settings
locust==2.22.0
pymongo==4.7.2
celery==5.2.7
redis==4.5.4